Adapted the code to work in batches, implemented testing and batch segmentation

+++ scripts/evaluation/image_datasets.py
@@ -1,26 +1,62 @@
-"""
-Given directories of labeled images, divide into sets for training, testing, and validation.
-Options: natural data augmentation, train/test/validation fractions, reduction in training size
-"""
-
 import pandas
 import os
 from glob import glob
+import evaluation_settings as s
 
 
-def create(label_dir, image_dir, output_dir, name, augment=False, augment_duration=0,
-           frac_train=0.7, frac_test=0.3, frac_validate=0.0, reduce_train=0.0,
-           label_pattern='*.png', image_ext='.jpg'):
+def create_all(working_dir, augment=False, augment_duration=0, reduce_train=0.0,
+               label_pattern='*.png', image_pattern='*.jpg', force=False):
+    """
+    Given directories of labeled images, divide into sets for training, testing, and validation.
+    Options: natural data augmentation, train/test/validation fractions, reduction in training size
+    :param working_dir: working directory of project
+    :param augment: augment the dataset by increasing the validity period of each label
+    :param augment_duration: duration of the validity period
+    :param reduce_train: reduce the fraction of training data by this fraction (the images left out are not used at all)
+    :param label_pattern: pattern for discovering label files in folder
+    :param image_pattern: pattern for discovering image files in folder
+    :return: path to csv file
+    """
 
+    # directory containing labeled images
+    label_dir = os.path.join(working_dir, s.stages[2])
+    # directory containing images (not labeled)
+    image_dir = os.path.join(working_dir, s.stages[1])
+    # where the dataset definition should be stored
+    output_dir = os.path.join(working_dir, s.stages[3])
+
+    datasets = []
+
+    for label_subdir in next(os.walk(label_dir))[1]:
+        camera, mode = label_subdir.split('_')
+        frac_train = s.datasets[mode]['frac_train']
+        frac_validate = s.datasets[mode]['frac_validate']
+        frac_test = s.datasets[mode]['frac_test']
+        for image_subdir in next(os.walk(image_dir))[1]:
+            camera2, multitime = image_subdir.split('_', 1)
+            if camera != camera2:
+                continue
+            else:
+                name = '_'.join([camera, mode, multitime])
+                datasets.append(create(label_dir=os.path.join(label_dir, label_subdir), image_dir=os.path.join(image_dir,image_subdir), output_dir=output_dir, name=name,
+                                       augment=augment, augment_duration=augment_duration, frac_train=frac_train,
+                                       frac_test=frac_test, frac_validate=frac_validate, reduce_train=reduce_train,
+                                       label_pattern=label_pattern, image_pattern=image_pattern))
+    return datasets
+
+def create(label_dir, image_dir, output_dir, name, augment=False, augment_duration=0, frac_train=0.7, frac_test=0.3,
+           frac_validate=0.0, reduce_train=0.0, label_pattern='*.png', image_pattern='*.jpg', force=False):
     # fetch list of label images
-    labels = glob(os.path.join(label_dir, label_pattern))
+    labels = glob(os.path.join(label_dir, 'labels', label_pattern))
+    if len(labels) == 0:
+        return
     labels.sort()
     labels_df = pandas.DataFrame(dict(label_path=labels))
     labels_df['time'] = labels_df['label_path'].apply(get_time_str)
 
     # fetch corresponding images
     if not augment:
-        images_all = glob(os.path.join(image_dir, '*' + image_ext))
+        images_all = glob(os.path.join(image_dir, image_pattern))
         images_all.sort()
         images_df = pandas.DataFrame(dict(image_path=images_all))
         images_df['time'] = images_df['image_path'].apply(get_time_str)
@@ -37,9 +73,9 @@ def create(label_dir, image_dir, output_dir, name, augment=False, augment_durati
 
     # determine numbers of train, test, and validation
     sample_counts = {
-        "train": frac_train*images_and_labels.shape[0]*(1-reduce_train),
-        "test": frac_test*images_and_labels.shape[0],
-        "validate": frac_validate*images_and_labels.shape[0]
+        "train": frac_train * images_and_labels.shape[0] * (1 - reduce_train),
+        "test": frac_test * images_and_labels.shape[0],
+        "validate": frac_validate * images_and_labels.shape[0]
     }
     images_and_labels['role'] = ''
     counter = 0
@@ -49,7 +85,8 @@ def create(label_dir, image_dir, output_dir, name, augment=False, augment_durati
 
     # Write to csv file
     csv_file = os.path.join(output_dir, name + '.csv')
-    images_and_labels.to_csv(csv_file)
+    if not os.path.exists(csv_file) or force:
+        images_and_labels.to_csv(csv_file)
 
     return csv_file
